[NUNCOTA-169] Fix race conditions in wsserver
There are two types of race condition in wsserver:

* a client is connected during server is closing: the client is not
notified because it is not in client list yet;
* when a client just connected but message processor is not properly
updated and may contains old value.

Fixes are:

First race is fixed by locking a crating of new connection and unlock only
after the client handler is put to the client list. The second race is
fixed by removing message processor interface and using simple process
message callback.

Also unit tests are refactored in a way the server is created for each
individual test.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>